Use ToPerformer in identify
